Improves app stability and UI
Addresses potential errors by checking if the widget is mounted before calling `context.loaderOverlay.hide()` in authentication and job creation screens.

Replaces `ButtonBar` with `OverflowBar` in the home screen for better layout flexibility.

Uses `withValues` instead of `withOpacity` for color manipulation in the `MiniJobCard` and `NotificationTile` to avoid potential issues.

Removes unused `_query` variable in the search screen.

Uses `loremPicsum` for the company logo to provide a placeholder image.